FIX: delete fitter objects on burst_search() and calc_fret() calls. Addresses issue #50.
The method calc_fret was testing for existence of attributes to be
deleted (including fitter attributes) using `attr in self` syntax.
Since fitter attributes are not items of the Data dict they were not
detected and deleted. This caused the display of stale FRET or S
histograms when several burst search were performed.

Thanks to Danielis Rutkauskas for finding and reporting this bug.